Move backfilling out of boot and into worker-thread
The boot process doesn't need to know about web-related
compromises. That can be handled by the process that
invokes bootWordPress().
